refactor: Update accelerate_config.yaml for DEEPSPEED configuration and mixed precision
Refactor the accelerate_config.yaml file to update the DEEPSPEED configuration and enable mixed precision. This includes modifying the distributed_type from MULTI_GPU to DEEPSPEED, and changing the value of mixed_precision from "no" to "bf16". Additionally, the downcast_bf16 option is set to "yes" for downcasting to bf16. The num_processes is also updated from 1 to 2.

Update train.py to load a different model and tokenizer

In train.py, the model_id is changed to "Gunulhona/Gemma-Ko-Merge" and the model and tokenizer are loaded using the AutoModelForCausalLM and AutoTokenizer classes. The previous code for loading the model and tokenizer using FastLanguageModel is commented out. The LoraConfig and LoHaConfig options for applying Lora to the model are also modified, but the specific changes are not mentioned in the commit message.